refactor(security): remove MFA card and replace with MFA method list for improved multi-factor authentication management
