+++ apps/web/messages/en.json
@@ -1756,8 +1756,6 @@
     "email-verified": "Email Verified",
     "verified": "Verified",
     "member-since": "Member Since",
-    "mfa": "Multi-Factor Authentication",
-    "mfa-description": "Manage your multi-factor authentication settings",
     "security-settings": "Security Settings",
     "security-settings-description": "Manage your password and account security",
     "linked-accounts": "Linked Accounts",
@@ -1804,7 +1802,6 @@
     "two-factor-auth": "Two-Factor Authentication",
     "two-factor-auth-description": "Add an extra layer of security to your account",
     "coming-soon": "Coming Soon",
-    "setup-2fa": "Set up 2FA",
     "active-sessions": "Active Sessions",
     "active-sessions-description": "View and manage your active login sessions",
     "view-sessions": "View Sessions",
@@ -1860,8 +1857,6 @@
     "password-management-description": "Change your password or reset it if forgotten",
     "two-factor-authentication": "Two-Factor Authentication",
     "two-factor-authentication-description": "Add an extra layer of security to your account",
-    "setup-2fa": "Setup 2FA",
-    "active-sessions-description": "Manage and monitor your active sessions ({count} active)",
     "security-recommendations-description": "We recommend completing these security measures",
     "use-strong-passwords-tip": "Use strong, unique passwords for all accounts",
     "avoid-public-wifi-tip": "Avoid accessing sensitive information on public WiFi",
