+++ apps/web/messages/vi.json
@@ -1768,8 +1768,6 @@
     "email-verified": "Email đã xác minh",
     "verified": "Đã xác minh",
     "member-since": "Thành viên từ",
-    "mfa": "Xác thực hai yếu tố",
-    "mfa-description": "Quản lý cài đặt xác thực hai yếu tố của bạn",
     "security-settings": "Cài đặt bảo mật",
     "security-settings-description": "Quản lý mật khẩu và bảo mật tài khoản của bạn",
     "linked-accounts": "Tài khoản liên kết",
@@ -1816,7 +1814,6 @@
     "two-factor-auth": "Xác thực hai yếu tố",
     "two-factor-auth-description": "Thêm lớp bảo mật bổ sung cho tài khoản của bạn",
     "coming-soon": "Sắp ra mắt",
-    "setup-2fa": "Thiết lập 2FA",
     "active-sessions": "Phiên hoạt động",
     "active-sessions-description": "Xem và quản lý các phiên đăng nhập đang hoạt động của bạn",
     "view-sessions": "Xem phiên",
@@ -1872,8 +1869,6 @@
     "password-management-description": "Thay đổi mật khẩu hoặc đặt lại nếu quên",
     "two-factor-authentication": "Xác thực hai yếu tố",
     "two-factor-authentication-description": "Thêm lớp bảo mật bổ sung cho tài khoản của bạn",
-    "setup-2fa": "Thiết lập 2FA",
-    "active-sessions-description": "Quản lý và theo dõi các phiên hoạt động của bạn ({count} đang hoạt động)",
     "security-recommendations-description": "Chúng tôi khuyên bạn nên hoàn thành các biện pháp bảo mật này",
     "use-strong-passwords-tip": "Sử dụng mật khẩu mạnh và duy nhất cho tất cả tài khoản",
     "avoid-public-wifi-tip": "Tránh truy cập thông tin nhạy cảm trên WiFi công cộng",
