+++ apps/web/src/app/[locale]/(dashboard)/settings/account/security/page.tsx
@@ -1,5 +1,4 @@
 import LinkedIdentitiesCard from './linked-identities-card';
-import MFACard from './mfa-card';
 import SecuritySettingsCard from './security-settings-card';
 import { createClient } from '@tuturuuu/supabase/next/server';
 
@@ -11,7 +10,6 @@ export default async function SecurityPage() {
 
   return (
     <div className="space-y-6">
-      <MFACard />
       <SecuritySettingsCard user={user} />
       <LinkedIdentitiesCard />
     </div>
